+++ src/constants.ts
@@ -1,7 +1,7 @@
 export namespace commmands {
     export const generateParameters = "arm.generateParameterFile";
     export const consolidateFiles = "arm.consolidateFiles";
-    export const extractParameter = "arm.extractParameter";
+    export const extractor = "arm.extractor";
 }
 
 export namespace config {